move render characters into different folder
